# routes/index.js
const express = require('express');
const router = express.Router();
const passport = require('passport');

// Authentication Routes
// Check if user is authenticated
router.get('/auth/status', /* #swagger.tags = ['Authentication'] */ (req, res) => {
  console.log('Status Check - Session:', req.session);
  console.log('Status Check - User:', req.user);
  console.log('Status Check - Is Authenticated:', req.isAuthenticated());
  
  if (req.isAuthenticated()) {
    res.status(200).json({
      isAuthenticated: true,
      user: req.user,
      sessionID: req.sessionID
    });
  } else {
    res.status(401).json({
      isAuthenticated: false,
      message: 'Not authenticated',
      session: req.session ? 'Session exists' : 'No session',
      sessionID: req.sessionID || 'No session ID'
    });
  }
});

// Redirects to GitHub for authentication
router.get('/login', 
  /* #swagger.ignore = true */
  (req, res, next) => {
    passport.authenticate('github', {
      successReturnToOrRedirect: '/api-docs',
      failureRedirect: '/api-docs',
      failureMessage: true
    })(req, res, next);
  });

// GitHub callback route
router.get('/auth/github/callback', /* #swagger.ignore = true */
  passport.authenticate('github', { 
    failureRedirect: '/api-docs',
    failureMessage: true,
    session: true
  }),
  (req, res) => { 
    console.log('Auth callback - User:', req.user);
    console.log('Auth callback - Session:', req.session);
    
      req.session.save((err) => {
      if (err) {
        console.error('Session save error:', err);
        return res.status(500).json({ error: 'Failed to save session' });
      }
      res.redirect('/api-docs');
    });
  }
);

// Logout route
router.get('/logout', /* #swagger.tags = ['Authentication'] */ (req, res, next) => {
    if (!req.isAuthenticated()) {
        return res.status(401).json({ message: 'You are not logged in' });
    }
    
    req.logout(function(err) {
        if (err) { 
            return res.status(500).json({ message: 'Error during logout' });
        }
        req.session.destroy(function(err) {
            if (err) {
                return res.status(500).json({ message: 'Error destroying session' });
            }
            res.status(200).json({ message: 'Logged out successfully' });
        });
    });
});
// --- End Auth Routes ---

router.use('/menuItems', require('./menuItems'));
router.use('/orders', require('./orders'));
router.use('/tables', /* #swagger.tags = ['Tables'] */ require('./tables'));
router.use('/employees', /* #swagger.tags = ['Employees'] */ require('./employees'));

module.exports = router;