# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Restaurant Management System API",
    "description": "Complete API for managing a restaurant's operations including menu items, orders, employees, and tables.\n\n    Key Features:\n    - Menu Management: Create and manage menu items with prices, categories, and availability\n    - Order Processing: Handle customer orders from creation to completion\n    - Employee Management: Track staff information, roles, and assignments\n    - Table Management: Monitor table status, capacity, and current orders\n    \n    Authentication Instructions:\n    1. To login: Click the Authorize button and use GitHub OAuth\n    2. Check authentication status: Use the /auth/status endpoint\n    3. To logout: Use the /logout endpoint\n    \n    Important Notes:\n    - POST, PUT, and DELETE operations require authentication\n    - All responses are in JSON format\n    - Dates should be provided in ISO 8601 format\n    - IDs are MongoDB ObjectIds\n    \n    Error Handling:\n    - 200: Success\n    - 201: Resource created\n    - 400: Invalid input\n    - 401: Not authenticated\n    - 404: Resource not found\n    - 500: Server error",
    "version": "1.0.0",
    "contact": {
      "name": "Restaurant API Support",
      "url": "https://github.com/WTHACD/CSE341-P"
    }
  },
  "host": "cse341-p-wfbq.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "githubAuth": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "/login",
      "description": "GitHub OAuth2 Authentication"
    }
  },
  "paths": {
    "/auth/status": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/menuItems/": {
      "get": {
        "tags": [
          "Menu Items"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Menu Items"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Pizza Margherita"
                },
                "description": {
                  "type": "string",
                  "example": "Classic pizza..."
                },
                "price": {
                  "type": "number",
                  "example": 12.5
                },
                "category": {
                  "type": "string",
                  "example": "Pizzas"
                },
                "stock": {
                  "type": "number",
                  "example": 50
                },
                "supplier": {
                  "type": "string",
                  "example": "Napoli Supplies"
                },
                "entryDate": {
                  "type": "string",
                  "example": "2025-09-26"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "type": "array",
                  "example": [
                    "60c72b2f9b1d8c001f8e4d2a"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "tableNumber": {
                  "type": "number",
                  "example": 5
                },
                "status": {
                  "type": "string",
                  "example": "received"
                },
                "notes": {
                  "type": "string",
                  "example": "Extra cheese"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/tables/": {
      "get": {
        "tags": [
          "Tables"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter tables by status",
            "required": false,
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "reserved"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Tables"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/employees/": {
      "get": {
        "tags": [
          "Employees"
        ],
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter employees by status",
            "required": false,
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          {
            "name": "role",
            "in": "query",
            "description": "Filter employees by role",
            "required": false,
            "type": "string",
            "enum": [
              "Waiter",
              "Cashier",
              "Cook",
              "Manager"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}